CAMEL-21843 - camel-kamelet - Move kamelet utils to corresponding components - Langchain4j
