+++ components/camel-ai/camel-langchain4j-chat/src/main/java/org/apache/camel/component/langchain4j/chat/huggingface/HugginFaceChatLanguageModelBuilder.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.kamelet.utils.langchain4j.chat;
+package org.apache.camel.component.langchain4j.chat.huggingface;
 
 import dev.langchain4j.model.huggingface.HuggingFaceLanguageModel;
