+++ components/camel-ai/camel-langchain4j-embeddings/src/main/java/org/apache/camel/component/langchain4j/embeddings/huggingface/HuggingFaceEmbeddingModelBuilder.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.kamelet.utils.langchain4j.embeddings;
+package org.apache.camel.component.langchain4j.embeddings.huggingface;
 
 import dev.langchain4j.model.huggingface.HuggingFaceEmbeddingModel;
